change everything back to float

+++ contraction.cu
@@ -247,7 +247,7 @@ double performContraction(std::vector<int> modeC, std::vector<int> modeA, std::v
 
     HANDLE_CUDA_ERROR(cudaEventRecord(stop, stream));
     HANDLE_CUDA_ERROR(cudaEventSynchronize(stop));
-    double milliseconds = 0;
+    float milliseconds = 0;
     HANDLE_CUDA_ERROR(cudaEventElapsedTime(&milliseconds, start, stop));
     // convert time to seconds
     milliseconds /= 1000;
@@ -281,7 +281,7 @@ double performContraction(std::vector<int> modeC, std::vector<int> modeA, std::v
     return milliseconds;
 }
 
-std::vector<double> run(std::vector<char> modeC, std::vector<char> modeA, std::vector<char> modeB, std::unordered_map<char, int64_t> extent)
+std::vector<float> run(std::vector<char> modeC, std::vector<char> modeA, std::vector<char> modeB, std::unordered_map<char, int64_t> extent)
 {
     // Convert char vectors to int vectors
     std::vector<int> modeC_int(modeC.begin(), modeC.end());
@@ -319,6 +319,6 @@ std::vector<double> run(std::vector<char> modeC, std::vector<char> modeA, std::v
     time5 = performContraction(modeC_int, modeA_int, modeB_int, extent_int, algo);
     HANDLE_CUDA_ERROR(cudaDeviceSynchronize());
 
-    std::vector<double> time = {time1, time2, time3, time4, time5};
+    std::vector<float> time = {time1, time2, time3, time4, time5};
     return time;
 }
